requirements.txt
langchain
langchain-community
langchain-text-splitters
chromadb
streamlit
python-dotenv
sentence-transformers
ollama
google-cloud-storage
google-cloud-firestore
google-auth
